# Conflicts:
#	src/LearnerData/SFA.DAS.LearnerData.Api/AppStart/AddServices.cs
#	src/LearnerData/SFA.DAS.LearnerData.UnitTests/Application/UpdateLearner/WhenHandlingUpdateLearnerCommand.cs
#	src/LearnerData/SFA.DAS.LearnerData/Application/UpdateLearner/UpdateLearnerCommandHandler.cs
#	src/LearnerData/SFA.DAS.LearnerData/Requests/CreateLearnerRequest.cs
#	src/LearnerData/SFA.DAS.LearnerData/Requests/UpdateLearnerRequest.cs

+++ src/ApprenticeApp/SFA.DAS.ApprenticeApp.MockApis/Program.cs
@@ -46,7 +46,7 @@ namespace SFA.DAS.ApprenticeApp.MockApis
                 var apprenticeship = Fake.CommitmentsApprenticeship;
                 var myApprenticeship = Fake.MyApprenticeship;
                 var provider = Fake.Provider;
-                provider.Ukprn = apprenticeship.ProviderId;
+                provider.Ukprn = (int)apprenticeship.ProviderId;
 
                 if (!args.Contains("!accounts", StringComparer.CurrentCultureIgnoreCase))
                 {
@@ -83,7 +83,7 @@ namespace SFA.DAS.ApprenticeApp.MockApis
                     _fakeCoursesInnerApi = new CoursesInnerApiMock(PortCoursesInnerApi, true)
                         .WithPing()
                         .WithStandardCourse(myApprenticeship.StandardUId, Fake.StandardApiResponse)
-                        .WithAnyStandardCourse(Fake.StandardApiResponse); 
+                        .WithAnyStandardCourse(Fake.StandardApiResponse);
                 }
 
                 Console.WriteLine($"Apprentice Id {apprentice.ApprenticeId}");
